[functorch] delete functorch/csrc/Constants.h
This file aliased dispatch keys. The original purpose was so that we
could change the dispatch keys in pytorch core without changing
functorch too much, but there's no need for the layer of indirection
anymore.

Also moved SINGLE_ARG to functorch/csrc/Macros.h, but that might need a
new home later.

ghstack-source-id: f27b105976af839f618a90aacd7e229fc1d37d35
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84639