Use presence of _symint in kernel name to generate symint sig or not
Something people found confusing was that whether or not a native::
signature would get SymInt or not in its type was based on the dispatch
key.  This changes it so that SymInt or not in type is based on whether
or not you have _symint in the name of the kernel or not.  This means
that even when we make operators support SymInt, you no longer have to
go and update all the preexisting definitions; instead, you now
selectively write _symint to opt individual kernels into SymInt support.

I then go and update a bunch of kernels that don't have proper SymInt
support to make use of this convention.  There is some hacking around
for view generation code.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: c42301c8280d41333bcd79722b49d2ce3039a2ba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84579